swapped interpreter and tracer dir names
